requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
discord==2.3.2
discord-py-interactions==4.4.1
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
importlib_metadata==8.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.20.0
